Migrate from Shakapacker/Webpack to Vite Rails
- Upgraded Node.js from 20.18.2 to 22.18.0 LTS
- Migrated package manager from Yarn v1.22.17 to pnpm 9.15.2
- Replaced Shakapacker 6.0.0 with vite_rails 3.0
- Configured Vite for React, TypeScript, and Rails integration
- Updated all CI/CD workflows and Dockerfiles
- Fixed JSX file extensions and dayjs import issues
- Configured proper aliasing for project structure
- Updated Rails views to use Vite helpers
- Significantly improved build performance and HMR speed

Breaking changes:
- Requires Node.js 22.x
- Uses pnpm instead of yarn
- Vite dev server runs on port 3036 instead of 3035
