+++ .github/workflows/validations.yml
@@ -55,22 +55,27 @@ jobs:
           bundle exec bundle-audit --update
           bundle exec ruby-audit check
       - name: Setup Node
-        uses: actions/setup-node@v1
+        uses: actions/setup-node@v4
         with:
-          node-version: 20.18.2
-      - name: Find yarn cache location
-        id: yarn-cache
-        run: echo "::set-output name=dir::$(yarn cache dir)"
+          node-version: 22.18.0
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v4
+        with:
+          version: 9.15.2
+      - name: Get pnpm store directory
+        id: pnpm-cache
+        shell: bash
+        run: echo "STORE_PATH=$(pnpm store path --silent)" >> $GITHUB_OUTPUT
       - name: JS package cache
         uses: actions/cache@v4
         with:
-          path: ${{ steps.yarn-cache.outputs.dir }}
-          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+          path: ${{ steps.pnpm-cache.outputs.STORE_PATH }}
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
           restore-keys: |
-            ${{ runner.os }}-yarn-
+            ${{ runner.os }}-pnpm-store-
       - name: Install packages
         run: |
-          yarn install --pure-lockfile
+          pnpm install --frozen-lockfile
       - name: Setup kernel for guard, increase watchers
         run: |
           echo fs.inotify.max_user_watches=524288 | sudo tee -a /etc/sysctl.conf && sudo sysctl -p
@@ -100,4 +105,4 @@ jobs:
           bundle exec rspec --exclude-pattern "spec/system/**/*_spec.rb"
       - name: Run Eslint
         run: |
-          npx eslint "./app/javascript/**/*.{ts,tsx,js,jsx,json}"
+          pnpm exec eslint "./app/javascript/**/*.{ts,tsx,js,jsx,json}"
