+++ .gitignore
@@ -41,6 +41,9 @@
 yarn-debug.log*
 .yarn-integrity
 
+# pnpm
+pnpm-debug.log*
+
 # Rails specific ignores
 /vendor/bundle
 
@@ -65,3 +68,11 @@ yarn-debug.log*
 # Claude Code Assistant
 .claude
 CLAUDE.md
+
+# Vite Ruby
+/public/vite*
+node_modules
+# Vite uses dotenv and suggests to ignore local-only env files. See
+# https://vitejs.dev/guide/env-and-mode.html#env-files
+*.local
+
