+++ .nvmrc
@@ -1,1 +1,1 @@
-20.18.2
+22.18.0
