+++ Procfile.dev
@@ -1,3 +1,3 @@
 web: bin/rails s -b 0.0.0.0 -p 3000
-webpacker: bin/webpacker-dev-server
+vite: bin/vite dev
 solidqueue: bundle exec rake solid_queue:start
