+++ app/views/layouts/application.html.erb
@@ -13,8 +13,10 @@
 
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-    <%= stylesheet_pack_tag "application" %>
-    <%= javascript_pack_tag "application" %>
+    <%= vite_client_tag %>
+    <%= vite_javascript_tag 'application' %>
+    <%= vite_stylesheet_tag 'application' %>
+
   </head>
   <body>
     <main class="min-h-screen bg-miru-white-400">
