+++ app/views/layouts/pdf.html.erb
@@ -4,7 +4,7 @@
     <meta http-equiv="content-type" content="text/html; charset=utf-8" />
     <script src="https://cdn.tailwindcss.com"></script>
     <%= stylesheet_link_tag "application" %>
-    <%= stylesheet_pack_tag "application" %>
+    <%= vite_stylesheet_tag 'application' %>
     <link href='https://fonts.googleapis.com/css?family=Roboto:400,300italic,300,400italic,500,500italic,700,700italic&amp;subset=all&device=swap&display=swap' rel='stylesheet' type='text/css'>
     <style>
     </style>
