+++ mise.toml
@@ -1,0 +1,2 @@
+[tools]
+node = "22.18.0"
