+++ package.json
@@ -11,12 +11,11 @@
     "@babel/runtime": "^7.21.0",
     "@fontsource/plus-jakarta-sans": "^4.5.0",
     "@headlessui/react": "^1.4.2",
-    "@heroicons/react": "^1.0.5",
+    "@heroicons/react": "v2",
     "@phosphor-icons/react": "^2.1.4",
     "@rails/actioncable": "^7.0.4",
     "@rails/activestorage": "^7.0.4",
     "@rails/ujs": "^7.0.4",
-    "@types/node": "^18.15.11",
     "@types/react": "^18.0.33",
     "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
@@ -93,6 +92,8 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.16.3",
+    "@types/node": "^18.15.11",
+    "@vitejs/plugin-react": "^5.0.0",
     "@webpack-cli/serve": "^2.0.1",
     "eslint": "^8.4.1",
     "eslint-config-prettier": "^8.3.0",
@@ -108,6 +109,9 @@
     "lint-staged": "^12.1.2",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
+    "vite": "^5.0.0",
+    "vite-plugin-ruby": "^5.1.0",
+    "vite-tsconfig-paths": "^5.1.4",
     "webpack-dev-server": "^4.13.1"
   },
   "lint-staged": {
@@ -132,7 +136,9 @@
   },
   "scripts": {
     "prepare": "husky install",
-    "lint": "eslint ./app/javascript"
+    "lint": "eslint ./app/javascript",
+    "dev": "pnpm exec webpack serve --config config/webpack/webpack.config.js",
+    "build": "pnpm exec webpack --config config/webpack/webpack.config.js"
   },
   "babel": {
     "presets": [
@@ -143,10 +149,10 @@
   "browserslist": [
     "defaults"
   ],
-  "resolutions": {
-    "**/react": "^18.2.0",
-    "**/react-dom": "^18.2.0"
+  "overrides": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "license": "MIT",
-  "packageManager": "yarn@1.22.17"
+  "packageManager": "pnpm@9.15.2"
 }
